Refactor rasterization args and tensor accessors
This PR Factors out common arguments from forward and backward 3DGS rasterization into a common class to reduce duplication, and converts backward rasterization to use tensor accessors rather than raw pointers.

The previous DeviceArgs structs in GaussianRasterizeBackward.cu and GaussianRasterizeForward.cu have been renamed to RasterizeBackwardArgs and RasterizeForwardArgs. There is now a RasterizeCommonArgs struct that each of the former structs composes (has-a). Much of the initialization and access to data that is common to both passes is now done in the new GaussianRasterize.cuh header. Other parts of the code are adjusted to be more uniform between the forward and backward passes in preparation for adding sparse sampling support to the backward pass.

---------

Signed-off-by: Mark Harris <mharris@nvidia.com>