Add NVTX convenience macros and use them in Gaussian Splatting
This adds new helper macros (FVDB_FUNC_RANGE()) that add NVTX ranges for the lifetime of a C++ function call. These can be visualized in NSight Systems timeline profiles to enable tracing the origin of kernels and CUDA API calls to specific locations in FVDB.

This PR also adds calls to the macros across the top levels of the Gaussian Splatting implementation.

---------

Signed-off-by: Mark Harris <mharris@nvidia.com>