Add Gaussian Splatting end-to-end benchmark
Fixes #763

This PR adds a benchmark.py in the 3d_gaussian_splatting project. The intent is to use this script for continuous benchmarking in CI once we have storage for checkpoints.

This loads checkpoints listed in a config file and runs the forward and backward 3dgs pipeline on them. It also has benchmarks for forward projection and rasterization individually.

In addition, the PR includes generate_benchmark_checkpoints.py, which can be used to run training on one or more datasets and save checkpoints at specified percentages. It updates benchmark_config.yaml with all generated checkpoints so the benchmark can easily be run on them.

The generate_benchmark_checkpoints.py also has an option --find-checkpoints-run-name=<str> which I mostly use during debugging. This skips training and just searches the directory results/<dataset>/<run_name>/checkpoints for checkpoint files and adds their paths to the config.yaml for configuring benchmark.py. Useful if you have an existing run but need to update the config yaml for benchmarking.

---------

Signed-off-by: Mark Harris <mharris@nvidia.com>