[Bugfix] Proper treatment of empty grids in topological operators
A few important fixes in the way topological ops work, especially in margin cases

Proper handling of empty grids as inputs to dilate/prune operators.
Proper handling of ordering of root nodes, to ensure the mapping from coord->index is the same for indexGrids with the same active voxel set, regardless of the path that their creation took (e.g. PointsToGrid vs TopoOps)

Signed-off-by: Eftychios Sifakis <esifakis@nvidia.com>