Add CODEOWNERS file (#17)
We need a CODEOWNERS file in order to have github auto-assign reviewers.
For now, all files are owned by @openvdb/fvdb-dev . In the future if we
have subsets of the code with more specific owners, we can add those
teams and matchers to `CODEOWNERS` to get more specific automatic review
assignments.

Signed-off-by: Mark Harris <mharris@nvidia.com>