Refactor common Gaussian rasterization grid/blockDim code and rename macros (#134)
The same gridDim and blockDim computation is used in the forward and
backward rasterization code so we can move these into the common args.
In addition, this also renames the macros for calling various templated
channel sizes. The new macro names no longer start with two underscores
(which is a prefix reserved by the compiler) and include the device type
in order to distinguish devices for future mGPU support.

---------

Signed-off-by: Matthew Cong <mcong@nvidia.com>