Add support for mGPU fused SSIM (#144)
This PR adds mGPU support for fused SSIM.

1. Adds the previously vendored single GPU fused SSIM kernels into the
codebase. This enables the addition of the mGPU + Unified Memory
variant, and fixes a small issue in the single GPU variant where the
stream was not specified. The existing copyright of the vendored code is
preserved with the OpenVDB copyright appended afterwards.
2. Defines the custom PyTorch operators `_fused_ssim` and
`_fused_ssim_backward` and their appropriate schemas. Note that the
names here are chosen (leading underscore, no trailing underscore) to be
consistent with existing PyTorch fused operator schemas.
3. Provides/binds the aforementioned fused SSIM kernels to these
operators for both the CUDA and PrivateUse1 backend.

As a result, PyTorch will dynamically dispatch to the CUDA or
PrivateUse1 version of the operator (just like any other PyTorch
operator) based on the selected devices. Furthermore, all the tests in
test_metrics.py pass without modification.

---------

Signed-off-by: Matthew Cong <mcong@nvidia.com>