[CI] Fix GPU instances not cleaning up their elastic storage (#154)
Switch cu12* tests to use the AMI for GPU instances that will cleanup
elastic storage on termination. This was costing more than necessary.

This also includes a fix to the conda based unit tests that have changed
from using `micromamba activate` to `conda activate` and is failing
tests which, based on looking at the micromamba PR, should not have
happened.

---------

Signed-off-by: Jonathan Swartz <jonathan@jswartz.info>