Docker workdir fixes (#152)
This change modifies the docker build so that it no longer uses the
parent directory, which is a holdover from the
previous OpenVDB repo.

It also uses a more tailored approach towards the mounting of local
directories.

Builds within the container will not affect build directories in the
local directory outside the container.

---------

Signed-off-by: Christopher Horvath <chorvath@nvidia.com>
Signed-off-by: Christopher Horvath <blackencino+git@gmail.com>
Co-authored-by: Jonathan Swartz <jonathan@jswartz.info>