Fix ODR violation introduced by viewer bindings (#168)
ViewerBindings.cpp uses a `GaussianSplat3d::ProjectionType` as a default
argument which means that it needs a type caster defined since default
arguments are instantiated at import time. Since ViewerBindings.cpp
wasn't including the header file for the `ProjectionType` typecaster, it
was using the default no-op typecaster provided by pybind11 in it's
translation unit hence the ODR violation warning.

The changes to the typecaster body itself are to make it more compliant
with pybind11's recommendations, including the use of the
`PYBIND11_TYPE_CASTER` macro. This fixes the second part of the ODR
warning which says that the fields in the defined type caster are
different from what pybind11 expects.

Signed-off-by: Matthew Cong <mcong@nvidia.com>