Move viz_edge_networ into viz module and split into multiple files (#186)
viz_edge_network is an attribute that we use for visualization of grid
edges. I think it belongs in the viz module as a standalone function. I
moved it there and renamed it to `grid_edge_network` and
`gridbatch_edge_network` which are better names. I also split the viz
module into files since viewer.py was getting big

---------

Signed-off-by: Francis Williams <francis@fwilliams.info>