Expose ProjectionType as a bound class instead of a typecaster (#185)
We've been running into ODR issues with
`GaussianSplat3d.ProjectionType`. See #167 , #168 , and #182. This PR
proposes a long-term solution by adding Python bindings for the
`ProjectionType` enum. This is preferable from a programmatic standpoint
because it avoids the pitfalls of type casters (e.g. their need to be
included consistently in all TUs, the allocation upon switching from
Python to C++ and vice versa) and it promotes stricter cross-language
type-safety. Furthermore, on the C++ side, `ProjectionType` is nested
within `GaussianSplat3d` and it makes sense from a consistency
standpoint to mimic that on the Python side as well. (Currently,
`GaussianSplat3d` is a bound class while it's nested `ProjectionType` is
a opaque implicit type caster.)

Note that this is breaking change from the Python standpoint. Existing
instances of `"perspective"` or `"orthographic"` will need to be changed
to their typed counterparts, but this is a relatively straightforward
change.

---------

Signed-off-by: Matthew Cong <mcong@nvidia.com>