[build] Add --threads to CUDA targets (#237)
Add --threads to CUDA targets based on number of CUDA archs we build
for. CMAKE_BUILD_PARALLEL_LEVEL is then adjusted by dividing by the
number of target CUDA archs.

I think this balances our resources better for faster build times when
building for multiple archs (like for wheels to distribute) but it has
no effect when building a wheel for a single arch (like tests or local
dev).

Signed-off-by: Jonathan Swartz <jonathan@jswartz.info>