Always use pytorch Scaled Dot Product Attention (#220)
We previously used cuDNN for Scaled Dot Product Attention on GPUs with
CC ?> 9.0. However we only test on CC 8.9 in CI so this path was
untested, and I found that it ~fails~ crashes on blackwell. Since
PyTorch implementation works on all GPUs we care about now, let's remove
that code path.

Signed-off-by: Mark Harris <mharris@nvidia.com>