Add mGPU Gaussian splatting operators (#230)
This PR adds multiGPU support for Gaussian splatting.
- The existing single GPU implementation are mostly unchanged aside from
the addition to chunk indexing offsets in order to re-use the code for
multiGPU. This is demonstrated by the single GPU tests passing without
modification.
- Switching from `atomicAdd` vs `atomicAdd_system` for the single GPU
case adds approximately 200 microseconds per iteration so it's
relatively negligible. Hence, we use `atomicAdd_system` for the single
GPU case as well in order to maximize code re-use.
- The `deviceOffsetAndCount` function has been renamed to `deviceChunk`.
Furthermore, an aligned variant, `deviceAlignedChunk`, has been
introduced. In addition, several places in the mGPU U-Net operators that
were computing the chunks manually have been changed to use the
`deviceChunk` function for consistency.
- GaussianRasterize operators have switched from using an explicitly
separated block size of `(C, H, W)` to an equivalent packed block size
of `(C * H * W, 1, 1)` with the per dimension index computation being
done within the kernel. By collapsing the block index to a linear index,
we simplify load balancing/chunking across GPUs especially for the case
where the number of GPUs is not a power of two.

---------

Signed-off-by: Matthew Cong <mcong@nvidia.com>