Dense API changed to Channels Major/Minor (#262)
Based on a more nuanced understanding of the convolution API, it is more
correct and less confusing to keep the spatial dimensions unscrambled in
conversion to and from dense tensors. The only variation we may need is
for the channels to be in the major axes vs. the minor axes. As a
shorthand, the two basic orderings are:

(B, X, Y, Z, C*) -> Channels Minor (C minor)  
(B, C*, X, Y, Z)  -> Channels Major (C major)

C minor is the "default" and aligns more closely with what fVDB expects
based on the layout of jagged feature tensors.

C major is the ordering expected by the convolution API, and aligns with
what torch.nn.Conv3d and torch.nn.functional.conv3d expect.

These replace the former "XYZC" and "CZYX" naming conventions.

---------

Signed-off-by: Christopher Horvath <chorvath@nvidia.com>