Switch from host to device synchronization for mGPU operators (#265)
In the existing code, each GPU synchronizes with the host thread at the
end of the mGPU operator. This conservatively ensures that the kernels
launched on each GPU for a given operator would complete before the next
set of kernels were launched for the subsequent operator.

This PR revises the synchronization to use inter-GPU synchronization via
CUDA events and achieve the same synchronization guarantees while also
bypassing the host. This enables mGPU operators to be queued up
asynchronously wrt to the host and improves throughput.

---------

Signed-off-by: Matthew Cong <mcong@nvidia.com>