GridBatch API Docs Updates (#281)
This is a large update to the docstrings in `grid_batch.py` based
largely on docstring changes in `grid.py`. It includes a couple of minor
fixes to docstrings `grid.py` and it removes all the docstrings written
into the `GridBatchBindings.cpp` because none of those get exposed to
the end user in the current organization.

---------

Signed-off-by: Jonathan Swartz <jonathan@jswartz.info>