Use smaller runtime images for non-build CI steps (#320)
Our C++ tests and documentation tests don't run any build steps so the
cudnn-devel image pulls in more than we need when the cudnn-runtime
image would be sufficient. This saves about a minute of job time and
about 5GB in bandwidth for each (uncached) pull.

We can't apply the same optimization to the PyTests because we build
wheels to test against during that job.

---------

Signed-off-by: Matthew Cong <mcong@nvidia.com>