Fix GHA error in (currently unused) nightly workflows (#329)
To disable these two nightly workflows, the entire "on:" section was
commented out. That is an error since a GHA workflow must have an "on:".
This PR just uncomments "on:" and "workflow_dispatch".

(I was seeing daily errors from these on my fork because I had actions
enabled. Not sure why they don't cause errors on the main repo.)

Signed-off-by: Mark Harris <mharris@nvidia.com>