Fix potential oversubscription of nvcc * cmake parallel threads (#351)
@harrism and I discovered that in high arch compilation situations and
high RAM machines, it's possible to get the logic that tries to spread
thread parallelism between nvcc and the cmake parallel job execution,
that you can end up with a value of nvcc threads * cmake threads that is
greater than your available thread count. Adding logic to cap this to
the thread count of the machine.

Signed-off-by: Jonathan Swartz <jonathan@jswartz.info>