Hot fixes for Jcat errors (#352)
Jcat was returning the C++ backend JaggedTensor object.

This, oddly, was not failing in many cases because most code paths
immediately call ".jdata", which worked on both the python wrapper and
the C++ backend, so we weren't seeing this fail.

I added additional tests for this.

I removed the module-level functions and added all of the creation functions directly onto JaggedTensor.
I additionally consolidated concatenation into from_cat on GridBatch and JaggedTensor.

I left behind jcat and the newly created gcat for convenience, but we should prefer the class methods.

---------

Signed-off-by: Christopher Horvath <chorvath@nvidia.com>