Bindings and Pytests for MCMC functions (#394)
Closes #376 

This PR adds the MCMC relocation and add noise methods to the
`GaussianSplat3D` class and adds Python bindings for them.

Additionally, the original relocation kernel was operating on plain
opacities and scales, which means calling code would have to convert
stored logit opacities / log scales before and after calling the
function. In this PR we move the conversion (sigmoid/logit and exp/log)
to and from inside the kernel to save extra torch tensor operations
(separate kernels). This makes GaussianRelocation more consistent with
other functions in `GaussianSplat3D` which expect logitOpacities and
logScales. We also move clamping of the opacity to the range
[minOpacity, 1 - FLT_EPS] inside the kernel. Also reorders the scales
and opacities parameters to match other functions (scales first).

---------

Signed-off-by: Mark Harris <mharris@nvidia.com>